Fix testsuite for windows.
Seems that dist zilla expects path in unix format only.  Giving path
in the correct format for the system will give an error (at least for
the Build add_files).  For the path of slurp_file, it seems that it
will accept windows paths but will convert them anyway if they are
unix paths so drop File::Spec there too.  Better to be consistent.